finished resource drops
create a prefab variant of the prefab base and then configure as needed. the resource node is placed in the room (for example a tree is a resource node and it drops wood when hit.) the drops need to have the code on the child changed to change its value.